refactor: Move sorting to internal/
